AERtypes, Body Parts! some fixes
Action Events now return structs, which can be used in the future if we
ever need return values.
A List of Body Parts was added to Body. Body Parts can queried for Type.
Which can  be used at the AbilityStore to limit certain Abilities. See
AbilityStore "HumanKeyMove" Function for example