Body parts and body part connections
So alot of this code was made and designed for the future. Atm the
framework is being put in to allow for coli of body parts. But also the
framework being set for the ability to have body parts attached to
multiple other parts attached to multiple other parts.... as well as a
way to send infomation and energy from part to  part (and call update
and etc). Future plans to allow Drawing (atm a lil hicup) as well as
break and dismembering parts (although thats almost done as well)