BodyPartDesigner
Big Changes!
Ability to use the bpDesigner to design a part shape, give it suture
spots, and register it for abilities
Saves into the BodyPartStore for use later in designing a character
More Nami functions, check NodeManipular.cs to the right to see new
features
BodyPartSaveLoader saves a hardcopy version of the created bodyparts so
they'll persist to each program use :D atm it is incomplete, but it
works. It is missing the functionality to turn a hashtable into a
string. Next step is to add that functionality to nami