GameFix --- Bank Framework
Change in the enum ObjectSpecificType didnt match the default values in
the AEstore. Fixed! Player can now move around

Also pushing the framework for the bank, so yu might have to include
alot of files. They are mostly empty to be honest. I found a way to
circumvent the bank system so we could focus on getting abilities and
bodyparts full functional first with custom made characters. With
abilities being so big, its best to get that done for the demo, people
cant tell we handmade the parts via non-intutive methods